pyproject.toml
[project]
name = "20250918-python-uv-byo-weather-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.14.1",
    "azure-identity>=1.19.0",
    "starlette>=0.41.0",
    "uvicorn>=0.34.0",
]